refactor : State의 상태 판단 로직 수정
