+++ src/main/java/domain/game/Hand.java
@@ -8,9 +8,6 @@ import java.util.stream.Collectors;
 
 public final class Hand {
 
-    public static final GamePoint BLACKJACK_POINT = GamePoint.of(21);
-    public static final int BLACKJACK_CARD_COUNT = 2;
-
     private final List<Card> cards;
     private final GamePoint gamePoint;
 
@@ -28,6 +25,10 @@ public final class Hand {
         return create(cards);
     }
 
+    public int cardSize() {
+        return cards.size();
+    }
+
     public GamePoint getGamePoint() {
         return gamePoint;
     }
@@ -36,9 +37,5 @@ public final class Hand {
         return cards.stream()
                 .map(Card::of)
                 .collect(Collectors.toUnmodifiableList());
-    }
-
-    public int cardSize() {
-        return cards.size();
     }
 }
