+++ src/main/java/domain/state/Hit.java
@@ -22,4 +22,9 @@ public final class Hit extends Running {
     public double calculateProfit(final double bet, final State dealerState) {
         throw new IllegalStateException("게임 중간에 수익률을 계산할 수 없습니다.");
     }
+
+    @Override
+    public boolean isBust() {
+        return hand.getGamePoint().isSameAs(BUST_POINT);
+    }
 }
