+++ src/main/java/domain/state/Running.java
@@ -1,8 +1,11 @@
 package domain.state;
 
+import domain.game.GamePoint;
 import domain.game.Hand;
 
 public abstract class Running extends State {
+
+    public static final GamePoint BUST_POINT = GamePoint.of(0);
 
     protected Running(final Hand hand) {
         super(hand);
@@ -12,4 +15,15 @@ public abstract class Running extends State {
     public final State stay() {
         return new Stay(hand);
     }
+
+    @Override
+    public boolean isStay() {
+        return false;
+    }
+
+    @Override
+    public boolean isBlackjack() {
+        return false;
+    }
+
 }
