+++ src/main/java/domain/state/State.java
@@ -8,10 +8,6 @@ import java.util.List;
 
 public abstract class State {
 
-    public static final GamePoint BUST_POINT = GamePoint.of(0);
-    public static final GamePoint BLACKJACK_POINT = GamePoint.of(21);
-    public static final int BLACKJACK_CARD_COUNT = 2;
-
     protected Hand hand;
 
     protected State(final Hand hand) {
@@ -32,16 +28,7 @@ public abstract class State {
         return hand.getGamePoint();
     }
 
-    public boolean isStay() {
-        return this instanceof Stay;
-    }
-
-    public boolean isBlackjack() {
-        return hand.cardSize() == BLACKJACK_CARD_COUNT &&
-                hand.getGamePoint().isSameAs(BLACKJACK_POINT);
-    }
-
-    public boolean isBust() {
-        return hand.getGamePoint().isSameAs(BUST_POINT);
-    }
+    public abstract boolean isStay();
+    public abstract boolean isBlackjack();
+    public abstract boolean isBust();
 }
